Create module validations
